updating website folder, removing uneeded files
