+++ buildspec.yml
@@ -28,8 +28,7 @@ phases:
       - echo Setting S3 bucket name...
       - ACCOUNT_ID=$(aws sts get-caller-identity --query Account --output text)
       - BUCKET_NAME="minimal-informer-codepipeline-artifacts-${ACCOUNT_ID}-${AWS_REGION}"
-      # - echo Uploading 'init' folder to $BUCKET_NAME...
-      # - aws s3 cp ./website_layout s3://$BUCKET_NAME/StaticWebsite/ --recursive
+      - aws s3 sync ./website_layout s3://$BUCKET_NAME/StaticWebsite/ --delete
 
       # Create the zip folder for lambda layer
       - echo "Zipping the Lambda Layer Folder with the npm install"
