+++ website_layout/index.html
@@ -20,7 +20,7 @@
             <div class = "world-card">
                 <a href="wold-news.html">
                     <div class = "card-image-wrapper">
-                        <img src = "images\world_news_img.png">
+                        <img src = "website_layout\images\world_news_img.png">
                     </div>
                     <div class = "card-content">
                         <h2 class = "card-title">World News</h2>
@@ -33,7 +33,7 @@
             <div class = "tech-card">
                 <a href = "tech-news.html">
                     <div class = "card-image-wrapper">
-                        <img src = "images\kling_klang_img.png">
+                        <img src = "website_layout\images\kling_klang_img.png">
                     </div>
                     <div class = "card-content">
                         <h2 class = "card-title">Tech News</h2>
@@ -46,7 +46,7 @@
             <div class = "travel-card">
                 <a href="travel-news.html">
                     <div class = "card-image-wrapper">
-                        <img src = "images\globe_img.png">
+                        <img src = "website_layout\images\globe_img.png">
                     </div>
                     <div class = "card-content">
                         <h2 class = "card-title">Travel News</h2>
