+++ website_layout/travel-news.html
@@ -1,0 +1,39 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <link rel="stylesheet" href="style.css">
+    <title>Advisory Scraper</title>
+</head>
+<body>
+    <div id="page-container">
+        <header class = header>
+            <div class = Header-Wrapper>
+                <h1>The Minimal Informer</h1>
+                <p>World events, Tech news, Travel, all in one place</p>
+            </div>
+        </header>
+        <div class = World-News-Buttons>
+            
+            <button class = "Advisory" id = "advisory_news_button">
+                Advisory
+            </button>
+            <button class = "Vacation" id = "vacation_news_button">
+                Germany
+            </button>
+            <button class = "Transportation" id = "transportation_news_button">
+               Japan
+            </button>
+        </div>
+        <div id="table-output-travel"></div>
+        <script type="module" src="website-write.js"></script>
+        <footer>
+            <a href="index.html">
+                <button class = back-button>
+                    Home Page
+                </button>
+            </a>
+        </footer>
+    </div>
+</body>
+
+</html>
